Refactor libs (group them using folders)
