+++ src/lib/auth/magicClient.ts
